+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Locations display as `{code} {name}` and thus should be sorted by code, name. Refs UICIRC-703.
 * Add RTL/Jest testing for `PatronNoticeDetail` component in `settings/PatronNotices`. Refs UICIRC-643.
 * Title level request setting cannot be disabled when there is an active title level request in the system. Refs UICIRC-708.
+* Add possible for resizing panel in circulation rules editor. Refs UICIRC-709.
 
 ## [6.0.0] (https://github.com/folio-org/ui-circulation/tree/v6.0.0) (2021-09-30)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v5.1.1...v6.0.0)