+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Add RTL/Jest testing for `PatronNoticeDetail` component in `settings/PatronNotices`. Refs UICIRC-643.
 * Title level request setting cannot be disabled when there is an active title level request in the system. Refs UICIRC-708.
 * Add possible for resizing panel in circulation rules editor. Refs UICIRC-709.
+* Prevent deletion of policies if they're in use. Refs UICIRC-669.
 
 ## [6.0.0] (https://github.com/folio-org/ui-circulation/tree/v6.0.0) (2021-09-30)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v5.1.1...v6.0.0)