+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 * Add RTL/Jest testing for `PatronNoticeDetail` component in `settings/PatronNotices`. Refs UICIRC-643.
 * Title level request setting cannot be disabled when there is an active title level request in the system. Refs UICIRC-708.
 * Add possible for resizing panel in circulation rules editor. Refs UICIRC-709.
+* Use constants instead of hardcoded value for query limits. Refs UICIRC-724.
+* Provide read-only access to loan policies. Refs UICIRC-693.
 * Prevent deletion of policies if they're in use. Refs UICIRC-669.
 
 ## [6.0.0] (https://github.com/folio-org/ui-circulation/tree/v6.0.0) (2021-09-30)